Add low-power jumper detection for a time-delay
This CL adds the basis for checking a jumper in a low-power way that can
be used to set a time-delay for the beeping for maximum sneakiness.

Right now the code doesn't really do much with that value besides beeping
more to let you know the jumper was disconnected.  In a future patch, I'll
impliment a time-delay before the first beeps start when the jumper is cut.

Signed-off-by: Charlie Mooney <cmooney3@gmail.com>