Adding in a pythong script to test the devices
This test works by streaming in audio data from a microphone put right next to
the beeper.  It computes a spectrogram on that data and picks out the 4khz tone
that the buzzer creates, nothing the time and duration.  The duration is a little
rough (since the nature of a spectrogram forces you to use windows of time) but
it's close enough that you can definitely pick out the beeps.

The data is all stored on the hdd as a pickled file.  Each beep is one dict with
several fields indicating:
  start/end time
  the start/end frequencies (for debugging)
  the start/end signal strengths (for debugging)

Next, I'll write some tools to process this data and tell you if it's working
as expected

Signed-off-by: Charlie Mooney <cmooney3@gmail.com>